Factor out tooltip into hook
